chat working fine, this commit before changing backend path api/v1

+++ back-end/src/main/java/tn/pfeconnect/pfeconnect/config/ApplicationAuditAware.java
@@ -9,9 +9,9 @@ import org.springframework.security.core.context.SecurityContextHolder;
 
 import java.util.Optional;
 
-public class ApplicationAuditAware implements AuditorAware<Long> {
+public class ApplicationAuditAware implements AuditorAware<Integer> {
     @Override
-    public Optional<Long> getCurrentAuditor() {
+    public Optional<Integer> getCurrentAuditor() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
         if (authentication == null ||
